新品首页Fragment的实现

+++ app/src/main/java/ucai/cn/day_filicenter/bean/NewGoodBean.java
@@ -44,7 +44,7 @@ public class NewGoodBean {
     private String colorName;
     private String colorCode;
     private String colorUrl;
-    private int addTime;
+    private long addTime;
     private boolean promote;
 
     private String shareUrl;
@@ -202,11 +202,11 @@ public class NewGoodBean {
         this.colorUrl = colorUrl;
     }
 
-    public int getAddTime() {
+    public long getAddTime() {
         return addTime;
     }
 
-    public void setAddTime(int addTime) {
+    public void setAddTime(long addTime) {
         this.addTime = addTime;
     }
